запись и чтение задач в localstorage

+++ sf_oopapp/src/utils.js
@@ -1,7 +1,10 @@
 import { User } from "./models/User";
 import { Task } from "./models/Task";
-import { tasks } from "./app";
+import { todoWrapper, btnWrapper } from "./app";
 import { appState } from "./app";
+import { trueUser, tasksArr, usersArr } from "./app";
+
+
 
 export const getFromStorage = function (key) {
   return JSON.parse(localStorage.getItem(key) || "[]");
@@ -14,37 +17,49 @@ export const addToStorage = function (obj, key) {
 };
 
 export const generateTestUser = function (User) {
-  localStorage.clear();
-  const testUser = new User("testUser", "test123");
+  //localStorage.clear();
+  const testUser = new User("testUser4", "test123");
   User.save(testUser);
 };
-
 
 // создаем кнопку submit и поле ввода
 
 export const addInp = function () {
-  const inputField = document.querySelector('.input-wrapper');
-  inputField.innerHTML = `<input id="inputField" type="text" placeholder = "введите таску" autofocus autocomplete = "off" > `;
-  document.querySelector('.btn-wrapper').innerHTML = `<button class ="btn-submit"> Submit </button> `;
+  const inputWrapper = document.querySelector('.input-wrapper');
+  inputWrapper.innerHTML = `<input id="inputField" type="text" placeholder = "введите задачу" autofocus autocomplete = "off" > `;
+  btnWrapper.innerHTML = `<button class ="btn-submit"> Submit </button> `;
   document.querySelector('.btn-submit').addEventListener('click', addToDo);
 }
 
 // создаем тудушки из поля ввода
 
 export const addToDo = function () {
- 
   let description = document.querySelector('#inputField').value;
-  document.querySelector('#inputField').value = "";
+  document.querySelector('#inputWrapper').value = "";
   if (description) {
-    document.querySelector('.todos-wrapper-ready').innerHTML += `<div class = "description"> ${description} </div> <br>`;
-    tasks.push(new Task(description));
-    appState._currentUser.toDo = tasks;
-    User.save(appState);
-    console.log(appState);
-  }
-  inputField.outerHTML = '';
-  document.querySelector('.btn-wrapper').innerHTML = `<button class="btn-add">+Add card</button>`; 
-  
+    todoWrapper.innerHTML += `<div class = "description"> ${description} </div> <br>`;
+    tasksArr.push(new Task(description));
+    trueUser.toDo = tasksArr;
+    for (let user of usersArr) {
+      if (trueUser.id == user.id) {
+        user = trueUser;
+        localStorage.clear();
+        localStorage.setItem(user.storageKey, JSON.stringify(usersArr));
+        break;
+      };
+    }; 
+  };
+  inputWrapper.outerHTML = '';
+  btnWrapper.innerHTML = `<button class="btn-add">+Add card</button>`; 
   document.querySelector('.btn-add').addEventListener('click', addInp);
 }
 
+//заполнение todo листа из localstorage
+export const fillHtmlList = () => {
+  todoWrapper.innerHTML = '';
+  if (tasksArr.length > 0) {
+    tasksArr.forEach((item, index) => {
+      todoWrapper.innerHTML += `<div class = "description"> ${item.description} </div> <br>`;
+    });
+  };
+};
